requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
Flask-SocketIO==5.3.7
python-dotenv==1.0.1
psycopg[binary]==3.3.3
gunicorn==22.0.0
eventlet==0.36.1
